pyproject.toml
[project]
name = "thatch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Yegor Kuznetsov" }
]
requires-python = ">=3.13"
dependencies = [
    "ipywidgets>=8.1.7",
    "pillow>=11.3.0",
    "psutil>=7.0.0",
    "numpy>=2.3.2",
    # "opencv-python>=4.11.0.86",
    # "polars>=1.31.0",
    "matplotlib>=3.10.3",
    "pyyaml>=6.0.2",
]

[build-system]
requires = ["uv_build>=0.8.0,<0.9"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "jupyter>=1.1.1",
]

[tool.ruff]
line-length = 88

[tool.ruff.format]
quote-style = "single"
docstring-code-format = false

[tool.pyright]
typeCheckingMode = "basic"
#reportCallIssue = false
#reportExplicitAny = false
#reportAny = false

[[tool.pyright.executionEnvironments]]
root = "tests"
reportCallIssue = "none"
reportAttributeAccessIssue = "none"

